refactor: live users panel buttons
The buttons on the right of each user in the Live Users panel are rendered conditionally - if the user is the current user, only the view profile button is seen and leads to the user dashboard page. Else, all other buttons are present (start chat, view profile and block), which functionalities will be done later.